requirements.txt
# Core dependencies
quart==0.18.4
quart-cors==0.6.0
python-dotenv==1.0.0

# AI and ML
torch==2.0.1
TTS==0.8.0
groq==0.4.1

# Database
supabase==1.0.3

# Authentication & Security
PyJWT==2.3.0
python-jose==3.3.0

# Utilities
emoji==2.2.0
numpy==1.21.5
asyncio==3.4.3
aiohttp==3.8.1

# Development & Testing
pytest==7.4.0
pytest-asyncio==0.21.1
black==23.3.0
flake8==6.0.0
mypy==1.3.0 